+++ main/dense64/test/org/ejml/alg/dense/linsol/GenericLinearSolverChecks_D64.java
@@ -136,14 +136,14 @@ public abstract class GenericLinearSolverChecks_D64 {
         assertTrue(solver.setA(A_good));
         double q_good;
         try {
-            q_good = solver.quality();
+            q_good = (double)solver.quality();
         } catch( IllegalArgumentException e ) {
             // quality is not supported
             return;
         }
 
         assertTrue(solver.setA(A_bad));
-        double q_bad = solver.quality();
+        double q_bad = (double)solver.quality();
 
         assertTrue(q_bad < q_good);
 
@@ -164,14 +164,14 @@ public abstract class GenericLinearSolverChecks_D64 {
         assertTrue(solver.setA(A));
         double q;
         try {
-            q = solver.quality();
+            q = (double)solver.quality();
         } catch( IllegalArgumentException e ) {
             // quality is not supported
             return;
         }
 
         assertTrue(solver.setA(Asmall));
-        double q_small = solver.quality();
+        double q_small = (double)solver.quality();
 
         assertEquals(q_small, q, UtilEjml.TEST_64F);
     }
@@ -262,7 +262,7 @@ public abstract class GenericLinearSolverChecks_D64 {
             double val = t[j];
 
             for( int i = 0; i < dof; i++ ) {
-                A.set(j,i,Math.pow(val,i));
+                A.set(j,i, Math.pow(val,i) );
             }
         }
